Main - refactor out the code responsible for parsing connection results
